requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
dotenv
celery
redis
flask
boto3
